# nooblyjs-application-wiki/backend/swagger/openapi.full.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NooblyJS Wiki Platform API",
    "description": "Unified REST API for the NooblyJS Wiki Platform (port 9101). Covers the Datasources module (`/api/*`), Wiki module (`/applications/wiki/api/*`) and Blueprint module (`/applications/blueprint/api/*`).",
    "version": "2.0.0",
    "contact": { "name": "NooblyJS Team" },
    "license": { "name": "ISC" }
  },
  "servers": [
    { "url": "/", "description": "Unified backend (port 9101)" }
  ],
  "components": {
    "securitySchemes": {
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie issued after login"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "token",
        "description": "Bearer token (24 hr). Obtain from POST /api/auth/token or the API tokens section of your profile."
      }
    },
...
  },
  "security": [
    { "SessionAuth": [] },
    { "BearerAuth": [] }
  ],
...
  "paths": {
...
    "/applications/wiki/api/comments": {
      "post": {
        "tags": ["Wiki - Social"],
        "summary": "Add a comment to a document",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["path", "spaceName", "text"], "properties": { "path": { "type": "string" }, "spaceName": { "type": "string" }, "text": { "type": "string" } } } } } },
        "responses": { "201": { "description": "Comment added" } }
      },
      "delete": {
        "tags": ["Wiki - Social"],
        "summary": "Delete a comment",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Deleted" } }
      }
    },
...
  }
}